Add tablegen backend wot emits structs for records
Given an input file containing no dags, emit a C++ file contains
declarations and definitions for classes and defs respectively.

This is a concrete demonstration that tablegen generates only records,
and a tablgen backend can convert those data into any format it wishes.
In LLVM's case it's also usually a bunch of C++, but C++ that the LLVM
compiler framework expects - and of much greater utility than this small
demo.

    $ llvm-tblgen --emit-cpp-structs < /dev/null
    // This code was generated by my fancy tablegen backend