[StubV] Add a really basic frame lowering impl
It doesn't deal with any non-zero stack size, so is useless, but at
least it won't needlessly halt compilation on simple tests.

The two big things a target needs to implement to handle the lowering of
stack frames are:

* emitPrologue
* emitEpilogue

These are responsible for emitting code at the beginning and end of
functions, to deal with whatever the target needs to do to set up the
stack frames and tear them down again.

This typically involves: calculating the size of the stack; adjusting
the stack pointer up/down by that amount; storing/restoring any
callee-saved registers to/from the stack. How exactly this is done is
dependent on the particular ABI/calling convention the target
uses, and may differ on a per-function basis.

Targets typically find themselves with quite complicated stack layouts
(handling alignment, padding, dynamic stack allocations, readjustment,
exception handling and the combinations of all of the above) and the
code can get quite gnarly and difficult to test. The stubv
implementation is currently so basic that any stack usage would crash
the compiler.