[StubV] Add ADDI instruction and impl copyPhysReg
Copying physical registers is an important target hook. It's expected to
be implemented and is called during Post-RA Pseudo Expansion to resolve
COPY instructions.

There's no dedicated "move" instruction in stubv/RISC-V but it's
typically done by the ADDI instruction, adding 0 to the source register.

This instruction has several aliases we've added to comply with GNU/LLVM
assemblers.