[StubV] Add an extremely basic instruction selector
The components of the most basic SelectionDAG instruction selector
possible are:

* Implementing LowerFormalArguments
* Implementing LowerReturn
* Adding patterns

LowerFormalArguments and LowerReturn must be implemented by a target,
and are responsible for lowering function arguments "into" the DAG, and
return values "out of" the DAG. This necessitates knowledge of the
ABI/calling convention.

This commit only supports the ILP32 ABI, which passes everything in up
to 8 32-bit GPRs (a0-a7) or spills to the stack. We can't select
load/stores so the stack path will eventually fail, but it felt weird to
leave it out. We obviously can't do var args or floats or vectors or
anything more complicated.

We also need a PseudoRET codegen instruction over the top of JALR to
encapsulate certain bits of information that LLVM needs to know
statically about "return instructions", which cannot be said about JALR
in general (e.g., isReturn=1). This pseudo is auto-expanded to JALR by
the StubVAsmPrinter pass using a method generated by TableGen so doesn't
need encoding info or an assembly mnemonic.