[StubV] Add initial framework for extensions
Includes the skeletons for the M and Zmmul extensions.

The correct usage of these subtarget predicates will ensure any
instructions guarded by the 'AssemblerPredicate's will only be visible
to the assembler - and any patterns guarded by the 'Predicate's only
available to the instruction selector - when the corresponding extension
is enabled.