[Draft] Add a custom ISD node for CTPOP and select it in C++
This commit imagines a CTPOP_MAGIC instruction on our hardware. We
custom lower/legalize CTPOP to the new CTPOP_MAGIC StubVISD node, and
manually select that node in C++ (as opposed to doing it in TableGen)

Note that we could continue to custom legalize (i8 ctpop) to ISD::CTPOP
and have custom lowering lower CTPOP to CTPOP_MAGIC. This is the more
principled way of doing it, on balance. Still, we're in control eh.

This is done purely to show how custom nodes are added, and how C++
selection is performed.