[Draft] Add a custom DAG combine for CTPOP_MAGIC
Now we pretend we have a second instruction, CTPOP_MAGIC_ADD, which can
fold an addition into the ctpop operation.

We continue to use the custom CTPOP_MAGIC ISD node as before, but now
expect that it has two operands. It's our job to maintain and assert any
invariants we impose on these nodes (e.g., that the second operand to
CTPOP_MAGIC must be a constant). We initially lower all CTPOP_MAGIC
nodes with an 'addend' of 0.

Since we're lowering via a custom node, we need to perform this combine
ourselves. We hook into PerformDAGCombine and try to find CTPOP_MAGIC
nodes whose input is an addition of a constant, and fold those into a
new CTPOP_MAGIC instruction with a non-zero addend.

We then use two patterns to match CTPOP_MAGIC with an addend of 0 to one
instruction, and a non-zero addend to CTPOP_MAGIC_ADD. A pattern with a
specific immediate value (zero) will be preferred over a non-specific
immediate value, so we don't expect to see a CTPOP_MAGIC_ADD instruction
with an addend of 0. This could be another target invariant, which we
could/should assert on.

Alternatively we could have two custom ISD nodes (with/without addend),
or have the two operations share an instruction. It's up to us, the ISA,
later backend passes, etc., to find the best solution to the problem.